refactor: domain 패키지의 객체가 view 패키지 객체에 의존하지 않도록 변경
