+++ src/main/java/step3/CarGenerator.java
@@ -1,5 +1,7 @@
 package step3;
 
+import step3.domain.Car;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
