+++ src/main/java/step3/Runner.java
@@ -1,5 +1,9 @@
 package step3;
 
+import step3.domain.RacingGame;
+import step3.view.InputView;
+import step3.view.ResultView;
+
 import static step3.util.StringUtil.splited;
 
 public class Runner {
@@ -9,7 +13,13 @@ public class Runner {
         int numberOfTry = InputView.numberOfTry();
 
         RacingGame game = new RacingGame(CarGenerator.generate(nameOfCars), numberOfTry);
-        game.play();
+
+        ResultView.printResultInfoMessage();
+        for (int round = 1; !game.isEnd(round); round++) {
+            game.playRound();
+            ResultView.printRoundResult(game.cars());
+        }
+        ResultView.printWinner(game.winnerNames());
     }
 
 }
