+++ src/test/java/step3/CarGeneratorTest.java
@@ -2,6 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Car;
 
 import java.util.List;
