+++ src/test/java/step3/CarTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.domain.Car;
 import step3.exception.CarNameExceedException;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
